refactor(teams): 팀 라이프사이클을 TeamOperations 순수 모듈로 분리
AgentTeamWatcher.ts 1076 → 398 LOC. create/pause/resume/pauseMember/
resumeMember/merge/remove 의 본문은 TeamOperations 위임. tmuxBin/tmuxEnv/
teamsDirFor 함수 주입 패턴으로 watcher (electron) 와 CLI (headless) 둘 다
재사용 가능하게.

공개 API 100% 호환 — preload.ts / main.ts IPC handler 변경 0.
configPathFor() 가 참조하던 미정의 this.workspacePath 필드 정리.
